% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_lignes.R
\name{echange_ligne}
\alias{echange_ligne}
\title{Échange deux lignes d'une matrice}
\usage{
echange_ligne(A, apartir, jusqua)
}
\arguments{
\item{A}{une matrice, possiblement une matrice augmentée pour la résoultion d'un SEL ou pour trouver l'inverse d'une matrice}

\item{apartir}{est le numéro de la première ligne}

\item{jusqua}{est le numéro de la deuxième ligne}
}
\value{
la matrice \code{A}, avec les lignes \code{apartir} et \code{jusqua} échangées
}
\description{
Cette opération correspond à échanger deux lignes d'une matrice
}
\seealso{
Other opérations élémentaires de lignes: \code{\link{mult_ligne}}
}
