% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_lignes.R
\name{mult_ligne}
\alias{mult_ligne}
\title{Additionne des multiples de deux lignes d'une matrice}
\usage{
mult_ligne(A, apartir, mult1, jusqua, mult2)
}
\arguments{
\item{A}{une matrice, possiblement une matrice augmentée pour la résolution d'un SEL ou pour trouver l'inverse d'une matrice}

\item{apartir}{est le numéro de la première ligne}

\item{mult1}{est le multiple de la première ligne}

\item{jusqua}{est le numéro de la deuxième ligne}

\item{mult1}{est le multiple de la deuxième ligne}
}
\value{
la matrice \code{A}, avec la ligne \code{jusqua} obtenue par l'addition
}
\description{
Cette opération correspond à multiplier la ligne i par une constante C1 et 
à lui additionner la ligne j multipliée par une constante C2
}
\seealso{
Other opérations élémentaires de lignes: \code{\link{echange_ligne}}
}
