% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_matrix.R
\name{print_matrix}
\alias{print_matrix}
\title{Affiche le code LaTeX de la matrice}
\usage{
print_matrix(A, crochets = TRUE, fractions = FALSE, digits = 0, ...)
}
\arguments{
\item{A}{une matrice}

\item{crochets}{logique; si \code{TRUE}, affichage de la matrice entre crochets, si \code{FALSE}, affichage de la matrice entre parenthèses.}

\item{fractions}{logique; si \code{TRUE}, essaie d'exprimer les nombres non-entiers sous forme rationnelle.}

\item{...}{arguments additionnels que nous envoyons à \code{xtable::xtableMatharray()}}

\item{digits;}{permet de choisir le nombre de chiffres à droite de la virgule lors de l'affichage}
}
\description{
Cette fonction permet l'affichage du code LaTeX d'une matrice. Elle 
permet de choisir l'affichage entre \code{crochets} ou parenthèses.
Nous pouvons déterminer le nombre de chiffres après la virgule avec
\code{digits} et d'afficher sous forme de \code{fractions} ou non.
}
\examples{
A <- matrix(c(2, 1, -1,
             -3, -1, 2,
             -2,  1, 2), 3, 3)
B <- matrix(c(8/3, -11/2, -3), 3, 1)

print_matrix(A)
print_matrix(A, digits = 2)
print_matrix(A, crochets = FALSE)
print_matrix(A/7, fractions = TRUE)
print_matrix(B, fractions = TRUE)

}
\author{
Marc-André Désautels
}
