% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_matrix_augm.R
\name{print_matrix_augm}
\alias{print_matrix_augm}
\title{Affiche le code LaTeX de la matrice augmentée}
\usage{
print_matrix_augm(A, sel = TRUE, crochets = TRUE, fractions = FALSE,
  digits = 0, ...)
}
\arguments{
\item{A}{une matrice}

\item{sel}{logique; si \code{TRUE}, on affiche la matrice augmentée d'un SEL, sinon on affiche la matrice augmentée de l'inversion d'une matrice.}

\item{crochets}{logique; si \code{TRUE}, affichage de la matrice entre crochets, si \code{FALSE}, affichage de la matrice entre parenthèses.}

\item{fractions}{logique; si \code{TRUE}, essaie d'exprimer les nombres non-entiers sous forme rationnelle.}

\item{...}{arguments additionnels que nous envoyons à \code{xtable::xtableMatharray()}}

\item{digits;}{permet de choisir le nombre de chiffres à droite de la virgule lors de l'affichage}
}
\description{
Cette fonction permet l'affichage du code LaTeX d'une matrice augmentée. Elle 
permet de choisir l'affichage entre \code{crochets} ou parenthèses.
Nous pouvons déterminer le nombre de chiffres après la virgule avec
\code{digits} et d'afficher sous forme de \code{fractions} ou non.
Le paramètre \code{sel} permet d'afficher la matrice augmentée d'un SEL ou la matrice
augmentée de l'inversion d'une matrice.
}
\examples{
A <- matrix(c(2, 1, -1,
             -3, -1, 2,
             -2,  1, 2), 3, 3)
B <- matrix(c(8/3, -11/2, -3), 3, 1)

print_matrix_augm(A)
print_matrix(A, digits = 2)
print_matrix(A, crochets = FALSE)
print_matrix(A/7, fractions = TRUE)
print_matrix(B, fractions = TRUE)

}
\author{
Marc-André Désautels
}
