% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMat.R
\name{printMat}
\alias{printMat}
\title{Affiche le code LaTeX de la matrice augmentée}
\usage{
printMat(A, digits = 2, fractions = FALSE, crochets = TRUE, sel = TRUE)
}
\arguments{
\item{A}{une matrice}

\item{fractions}{logique; si \code{TRUE}, essaie d'exprimer les nombres non-entiers sous forme rationnelle.}

\item{crochets}{logique; si \code{TRUE}, affichage de la matrice entre crochets, si \code{FALSE}, affichage de la matrice entre parenthèses.}

\item{sel}{logique; si \code{TRUE}, on affiche la matrice augmentée d'un SEL, sinon on affiche la matrice augmentée de l'inversion d'une matrice.}

\item{digits;}{permet de choisir le nombre de chiffres à droite de la virgule lors de l'affichage}
}
\description{
Cette fonction permet l'affichage du code LaTeX d'une matrice augmentée. Elle 
permet de choisir l'affichage entre \code{crochets} ou parenthèses.
Nous pouvons déterminer le nombre de chiffres après la virgule avec
\code{digits} et afficher sous forme de \code{fractions} ou non.
Le paramètre \code{sel} permet d'afficher la matrice augmentée d'un SEL ou la matrice
augmentée de l'inversion d'une matrice.
}
\examples{
A <- matrix(c(2, 1, -1,
             -3, -1, 2,
             -2,  1, 2), 3, 3)
B <- matrix(c(2, 1, -1, 4,
             -3, -1, 2, 6,
             -2,  1, 2, -4,
             0, 4, 7, -5), 4, 4)

printMat(A)
printMat(A/2, digits = 2)
printMat(A, crochets = FALSE)
printMat(A/7, fractions = TRUE)
printMat(B, sel = FALSE)
printMat(cbind(A,diag(ncol(A))), sel = FALSE)
}
\author{
Marc-André Désautels
}
